+++ src/mrfitty/combination_fit.py
@@ -22,11 +22,11 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 import collections
-import glob
 import itertools
 import logging
 from operator import attrgetter
 import os.path
+import time
 
 import bokeh.io
 import bokeh.models.layouts
@@ -36,43 +36,93 @@ import matplotlib
 matplotlib.use('pdf', warn=False, force=True)
 import matplotlib.pyplot as plt
 from matplotlib.backends.backend_pdf import PdfPages
-from matplotlib.offsetbox import AnchoredText
-import numpy as np
 import pandas as pd
 
-from mrfitty.base import PRM, ReferenceSpectrum, Spectrum, InterpolatedReferenceSpectraSet, SpectrumFit
-from mrfitty.linear_model import NonNegativeLinearRegression
+from mrfitty.base import InterpolatedReferenceSpectraSet, SpectrumFit
 
 
 class CombinationFitResults:
     """CombinationFitResults
 
     """
-    def __init__(self, spectrum, best_fit, sorted_component_count_fit_lists):
+    def __init__(self, spectrum, best_fit, component_count_fit_table):
         self.spectrum = spectrum
         self.best_fit = best_fit
-        self.sorted_component_count_fit_lists = sorted_component_count_fit_lists
+        self.component_count_fit_table = component_count_fit_table
 
 
 class AllCombinationFitTask:
-    def __init__(self, ls, reference_spectrum_list, unknown_spectrum_list, energy_range_builder, component_count_range=range(4)):
+    def __init__(self, ls, reference_spectrum_list, unknown_spectrum_list, energy_range_builder, best_fits_plot_limit, component_count_range=range(4)):
         self.ls = ls
         self.reference_spectrum_list = reference_spectrum_list
         self.unknown_spectrum_list = unknown_spectrum_list
         self.energy_range_builder = energy_range_builder
+        self.best_fits_plot_limit = best_fits_plot_limit
         self.component_count_range = component_count_range
+
         self.fit_table = collections.OrderedDict()
 
-    def fit_all(self):
+    def fit_all(self, plots_pdf_dp):
         log = logging.getLogger(name='fit_all')
         for unknown_spectrum in self.unknown_spectrum_list:
             log.debug('fitting %s', unknown_spectrum.file_name)
-            best_fit, sorted_component_count_fit_lists = self.fit(unknown_spectrum)
-            self.fit_table[unknown_spectrum] = CombinationFitResults(
+            t0 = time.time()
+            best_fit, fit_table = self.fit(unknown_spectrum)
+            t1 = time.time()
+            log.info('fit %s in %5.3fs', unknown_spectrum.file_name, t1-t0)
+
+            fit_results = CombinationFitResults(
                 spectrum=unknown_spectrum,
                 best_fit=best_fit,
-                sorted_component_count_fit_lists=sorted_component_count_fit_lists
+                component_count_fit_table=fit_table
             )
+            self.fit_table[unknown_spectrum] = fit_results
+
+            file_base_name, _ = os.path.splitext(os.path.basename(unknown_spectrum.file_name))
+            plots_pdf_fp = os.path.join(plots_pdf_dp, file_base_name + '_fit.pdf')
+            with PdfPages(plots_pdf_fp) as plot_file:
+                log.info('writing plots file {}'.format(plots_pdf_dp))
+                # create plot
+                log.info('plotting fit for %s', unknown_spectrum.file_name)
+
+                f = self.plot_fit(spectrum=unknown_spectrum, any_given_fit=fit_results.best_fit, title='Best Fit')
+                plot_file.savefig(f)
+
+                f = self.plot_stacked_fit(spectrum=unknown_spectrum, any_given_fit=fit_results.best_fit, title='Best Fit')
+                plot_file.savefig(f)
+
+                ordinal_list = ('1st', '2nd', '3rd', '4th', '5th', '6th', '7th', '8th', '9th', '10th')
+
+                # plot the best n-component fit
+                for n in sorted(fit_table.keys()):
+                    log.info('plotting %d-component fit for %s', n, unknown_spectrum.file_name)
+                    n_component_fit_results = fit_table[n]
+                    # TODO: move this loop to best subset selection
+                    # here only plot the best fit for each component count
+                    for i, fit in enumerate(n_component_fit_results):
+                        if hasattr(fit, 'median_Cp'):
+                            f = self.plot_fit(
+                                spectrum=unknown_spectrum,
+                                any_given_fit=fit,
+                                title='Best {}-Component Fit ({})'.format(n, i))
+                            plot_file.savefig(f)
+                        else:
+                            # plot just the best fit for n component(s)
+                            if i < self.best_fits_plot_limit:
+                                f = self.plot_fit(
+                                    spectrum=unknown_spectrum,
+                                    any_given_fit=n_component_fit_results[i],
+                                    title='{} Best {}-Component Fit'.format(ordinal_list[i], n))
+                                plot_file.savefig(f)
+                            else:
+                                break
+
+                f = self.plot_nss_path(spectrum=unknown_spectrum, fit_results=fit_results, title='NSS Path')
+                plot_file.savefig(f)
+
+            nss_path_plot_fp = os.path.join(plots_pdf_dp, file_base_name + '_nss_path.html')
+            self.bokeh_nss_path(spectrum=unknown_spectrum, fit_results=fit_results, output_fp=nss_path_plot_fp)
+
         return self.fit_table
 
     def fit(self, unknown_spectrum):
@@ -82,14 +132,15 @@ class AllCombinationFitTask:
             unknown_spectrum=unknown_spectrum,
             reference_set=self.reference_spectrum_list)
         # fit all combinations of reference_spectra
+        # all_counts_spectrum_fit_table looks like this:
+        #   { 1: [...list of 1-component fits sorted by NSS...],
+        #     2: [...list of 2-component fits sorted by NSS...],
+        #     ...
+        #   }
         all_counts_spectrum_fit_table = collections.defaultdict(list)
         for reference_spectra_combination in self.reference_combination_iter(self.component_count_range):
             log.debug('fitting to reference_spectra {}'.format(reference_spectra_combination))
 
-            #spectrum_fit = self.fit_unknown_spectrum_to_references(
-            #    unknown_spectrum=unknown_spectrum,
-            #    reference_spectra_combination=reference_spectra_combination
-            #)
             spectrum_fit = self.fit_references_to_unknown(
                 interpolated_reference_spectra=interpolated_reference_spectra,
                 reference_spectra_subset=reference_spectra_combination)
@@ -99,28 +150,40 @@ class AllCombinationFitTask:
             spectrum_fit_list.sort(key=attrgetter('nss'))
             # when there are many reference spectra the list of fits can get extremely long
             # and eat up all the memory
+            # keep only the top 100 fits for each component count
             if len(spectrum_fit_list) > 100:
                 spectrum_fit_list.pop()
 
-        all_counts_spectrum_fit_list = []
-        for reference_count, spectrum_fit_list in all_counts_spectrum_fit_table.items():
-            all_counts_spectrum_fit_list.extend(spectrum_fit_list)
-        best_fit, sorted_component_count_fits = self.sort_fits(unknown_spectrum, all_counts_spectrum_fit_list)
-        return best_fit, sorted_component_count_fits
+        """
+        TODO: this should go somewhere else
+        # calculate prediction error on the best fits to find indistinguishable fits
+        for component_count, fit_list in all_counts_spectrum_fit_table.items():
+            for f, fit in enumerate(fit_list):
+                prediction_error_list = self.calculate_prediction_error_list(fit)
+                fit.median_Cp = np.median(prediction_error_list)
+                fit.median_Cp_ci_lo, fit.median_Cp_ci_hi = scikits.bootstrap.ci(
+                    data=prediction_error_list,
+                    statfunction=np.median)
+                log.info(
+                    'fit %d has Cp %8.5f <-- %8.5f --> %8.5f',
+                    f, fit.median_Cp_ci_lo, fit.median_Cp, fit.median_Cp_ci_hi)
+                # in the first iteration fit_list[0] == fit
+                if fit_list[0].median_Cp_ci_hi < fit.median_Cp_ci_lo:
+                    log.info('median Cp confidence interval for fit %d does not overlap that of the best fit', f)
+                    break
+        """
+        best_fit = self.choose_best_component_count(all_counts_spectrum_fit_table)
+        return best_fit, all_counts_spectrum_fit_table
 
     def reference_combination_iter(self, component_count_range):
         for component_count in component_count_range:
             for reference_spectra_combination in itertools.combinations(self.reference_spectrum_list, component_count):
                 yield reference_spectra_combination
 
-    #@profile
     def fit_references_to_unknown(self, interpolated_reference_spectra, reference_spectra_subset):
         interpolated_reference_spectra_subset_df, unknown_spectrum_df = \
             interpolated_reference_spectra.get_reference_subset_and_unknown_df(
                 reference_list=reference_spectra_subset)
-
-        #self.log.debug('unknown_spectrum_df:\n%s', unknown_spectrum_df.head())
-        #self.log.debug('interpolated_reference_spectra_subset_df:\n%s', interpolated_reference_spectra_subset_df.head())
 
         lm = self.ls()
         lm.fit(interpolated_reference_spectra_subset_df.values, unknown_spectrum_df.norm.values)
@@ -135,112 +198,18 @@ class AllCombinationFitTask:
         )
         return spectrum_fit
 
-    #@profile
-    def fit_unknown_spectrum_to_references(self, unknown_spectrum, reference_spectra_combination):
-        log = logging.getLogger(name=self.__class__.__name__)
-        fit_energies, fit_energies_ndx = self.energy_range_builder.build_range(
-            unknown_spectrum,
-            reference_spectra_combination
-        )
-        # interpolate the reference spectra at the fit_energies
-        log.debug('fit energies.shape: %s', fit_energies.shape)
-        # reference_spectra_A_df is energies x components
-        unknown_spectrum_b = unknown_spectrum.data_df.loc[fit_energies_ndx, 'norm']
-        # log.debug('unknown_spectrum_b.shape: {}'.format(unknown_spectrum_b.shape))
-        log.debug('unknown_spectrum_b      : %s', unknown_spectrum_b)
-        # log.debug('reference combination  : {}'.format(reference_spectra_combination))
-        reference_spectra_A_column_list = []
-        # create an array to hold the interpolated reference spectra
-        data = np.zeros((fit_energies.shape[0], len(reference_spectra_combination)))
-        for i, rs in enumerate(reference_spectra_combination):
-            reference_spectra_A_column_list.append(rs.file_name)
-            data[:, i] = rs.interpolant(fit_energies)
-        # use the (time) index from the unknown spectrum for the interpolated reference spectra
-        # this is important because pandas Series and Dataframes will align on their
-        # indexes for most operations so for example calculating residuals can result
-        # in the wrong shape
-        reference_spectra_A_df = pd.DataFrame(
-            data=data, index=unknown_spectrum_b.index, columns=reference_spectra_A_column_list)
-        log.debug('reference_spectra_A_df columns: %s', reference_spectra_A_df.columns)
-
-        # it is important to label the columns in the order they were appended
-
-        ls = NonNegativeLinearRegression()
-        ls.fit(reference_spectra_A_df.values, unknown_spectrum_b)
-        reference_spectra_coef_x = ls.coef_
-
-        #reference_spectra_coef_x, residual, *extra = nnls(
-        #    reference_spectra_A_df.values,
-        #    unknown_spectrum_b
-        #)
-        #reference_spectra_coef_x, residual, rank, sigma = lstsq(
-        #    reference_spectra_A_df.values,
-        #    unknown_spectrum_b
-        #)
-        # log.debug('A        :\n{}'.format(reference_spectra_A_df))
-        # log.debug('coef     : {}'.format(reference_spectra_coef_x))
-        # log.debug('residual : {}'.format(residual))
-        # log.debug('solution :\n{}'.format(reference_spectra_A_df.dot(reference_spectra_coef_x)))
-        #if np.any(reference_spectra_coef_x < 0.0):
-        #    # this happens a lot with lstsq and is generally not a problem
-        #    # print('{}: least-squares fit has negative coefficients'.format(
-        #    #    unknown_spectrum_file_name
-        #    # ))
-        #    continue
-        ##else:
-        spectrum_fit = SpectrumFit(
-            interpolant_incident_energy=fit_energies,
-            reference_spectra_A_df=reference_spectra_A_df,
-            unknown_spectrum_b=unknown_spectrum_b,
-            reference_spectra_seq=reference_spectra_combination,
-            reference_spectra_coef_x=reference_spectra_coef_x
-        )
-        return spectrum_fit
-
-    def sort_fits(self, spectrum, spectrum_fit_list):
-        # sort all fits with the same component count
-        log = logging.getLogger(name=spectrum.file_name)
-        # compare the top fits for each component count
-        # component_count_fit_lists looks like this:
-        #   [ [list of 0-component fits], [list of 1-component fits], ..., [list of n-component fits] ]
-        component_count_fit_lists = [[] for component_count in self.component_count_range]
-        # append an extra empty list for 0-component fits even though there are none
-        # this allows component count to work as the list index
-        component_count_fit_lists.append([])
-        log.debug(
-            'creating one fit list for each component count in self.component_count_range: %s',
-            self.component_count_range)
-        log.debug('component_count_fit_lists: %s', component_count_fit_lists)
-
-        # populate the component count lists from a sorted list of all fits
-        for spectrum_fit in sorted(spectrum_fit_list, key=attrgetter('nss')):
-            component_count = len(spectrum_fit.reference_spectra_seq)
-            component_count_fit_lists[component_count].append(spectrum_fit)
-
-        for component_count, component_count_fit_list in enumerate(component_count_fit_lists):
-            if len(component_count_fit_list) > 0:
-                log.debug('best fit for %d component(s): %s', component_count, component_count_fit_list[0])
-            else:
-                log.debug('no fits for component count %s', component_count)
-
-        best_fit_for_component_count_list = [c[0] for c in component_count_fit_lists[1:]]
-        best_fit = self.choose_best_component_count(best_fit_for_component_count_list)
-        return best_fit, component_count_fit_lists[1:]
-        # choose the best fit from the top fits for all component counts
-        # skip the 0-component list since it is empty
-
-    def choose_best_component_count(self, best_fit_for_component_count_list):
+    def choose_best_component_count(self, all_counts_spectrum_fit_table):
         """
         Choose the best fit from the best fits for each component count.
-        :param best_fit_for_component_count_list:
-          a list of fits for component counts 0 to N; there is no 0-component fit
-        :return:
+        :param all_counts_spectrum_fit_table:
+          dictionary with component count keys and values list of spectrum fits in sorted order
+        :return: instance of SpectrumFit
         """
         log = logging.getLogger(name=self.__class__.__name__)
         best_fit = None
         previous_nss = 1.0
-        for component_count, component_count_fit_list in enumerate(best_fit_for_component_count_list):
-            best_fit_for_component_count = component_count_fit_list
+        for component_count in sorted(all_counts_spectrum_fit_table.keys()):
+            best_fit_for_component_count = all_counts_spectrum_fit_table[component_count][0]
             improvement = (previous_nss - best_fit_for_component_count.nss) / previous_nss
             log.debug('improvement: {:5.3f} for {}'.format(improvement, best_fit_for_component_count))
             if improvement < 0.10:
@@ -264,10 +233,10 @@ class AllCombinationFitTask:
                 table_file.write('\t')
                 table_file.write('{:5.3f}\t'.format(fit_results.best_fit.nss))
                 table_file.write('{:5.3f}'.format(fit_results.best_fit.residuals_contribution))
-                for row in fit_results.best_fit.reference_contribution_percent_sr.sort_values(ascending=False).iteritems():
+                for ref_name, ref_pct in fit_results.best_fit.reference_contribution_percent_sr.sort_values(ascending=False).items():
                     table_file.write('\t')
-                    table_file.write(row[0])
-                    table_file.write('\t{:5.3f}'.format(row[1]))
+                    table_file.write(ref_name)
+                    table_file.write('\t{:5.3f}'.format(ref_pct))
                 table_file.write('\n')
 
     def draw_plots_matplotlib(self, plots_pdf_file_path):
@@ -277,39 +246,142 @@ class AllCombinationFitTask:
             for spectrum, fit_results in self.fit_table.items():
                 log.info('plotting fit for {}'.format(spectrum.file_name))
 
-                longest_name_len = max([len(name) for name in fit_results.best_fit.reference_contribution_percent_sr.index])
-                # the format string should look like '{:N}{:5.2f}' where N is the length of the longest reference name
-                contribution_format_str = '{:' + str(longest_name_len + 4) + '}{:5.2f}'
-                contribution_desc_lines = []
-                fit_results.best_fit.reference_contribution_percent_sr.sort_values(ascending=False, inplace=True)
-                for name, value in fit_results.best_fit.reference_contribution_percent_sr.iteritems():
-                    contribution_desc_lines.append(contribution_format_str.format(name, value))
-                contribution_desc_lines.append(
-                    contribution_format_str.format('residual', fit_results.best_fit.residuals_contribution))
-                contribution_desc = '\n'.join(contribution_desc_lines)
-
-                f, ax = plt.subplots()
-                f.suptitle(spectrum.file_name)
-                log.info(fit_results.best_fit.fit_spectrum_b.shape)
-                ax.plot(fit_results.best_fit.interpolant_incident_energy, fit_results.best_fit.fit_spectrum_b)
-                log.info(fit_results.best_fit.residuals.shape)
-                ax.plot(fit_results.best_fit.interpolant_incident_energy, fit_results.best_fit.unknown_spectrum_b, '.')
-                ax.plot(fit_results.best_fit.interpolant_incident_energy, fit_results.best_fit.residuals)
-
-                at = AnchoredText(contribution_desc, loc=1, prop=dict(fontname='Monospace', size=10))
-                ax.add_artist(at)
-
-                #log.info('fit_results.best_fit.interpolant_incident_energy:\n{}'.format(
-                #    fit_results.best_fit.interpolant_incident_energy)
-                #)
-                #ax.vlines(x=[
-                #        fit_results.best_fit.interpolant_incident_energy.iloc[0],
-                #        fit_results.best_fit.interpolant_incident_energy.iloc[-1]
-                #    ],
-                #    ymin=fit_results.best_fit.unknown_spectrum_b.min(),
-                #    ymax=fit_results.best_fit.unknown_spectrum_b.max()
-                #)
+                f = self.plot_fit(spectrum, fit_results.best_fit)
                 plot_file.savefig(f)
+
+                # plot the best n-component fit
+                #for n_fit_list in fit_results:
+                # plot the best 2-component fit
+
+    def plot_nss_path(self, spectrum, fit_results, title):
+        log = logging.getLogger(name=self.__class__.__name__)
+
+        f, ax = plt.subplots()
+        f.suptitle(spectrum.file_name + '\n' + title)
+        for component_count in fit_results.component_count_fit_table.keys():
+            sorted_fits = fit_results.component_count_fit_table[component_count]
+            ax.plot(range(len(sorted_fits)), [spectrum_fit.nss for spectrum_fit in sorted_fits])
+
+        return f
+
+    def plot_fit(self, spectrum, any_given_fit, title):
+        log = logging.getLogger(name=self.__class__.__name__)
+
+        f, ax = plt.subplots()
+        f.suptitle(spectrum.file_name + '\n' + title)
+        log.info(any_given_fit.fit_spectrum_b.shape)
+
+        reference_contributions_percent_sr = any_given_fit.get_reference_contributions_sr()
+        reference_only_contributions_percent_sr = any_given_fit.get_reference_only_contributions_sr()
+        longest_name_len = max([len(name) for name in reference_contributions_percent_sr.index] + [len(spectrum.file_name)])
+        # the format string should look like '{:N}{:5.2f} ({:5.2f})' where N is the length of the longest reference name
+        reference_contribution_format_str = '{:' + str(longest_name_len + 4) + '}{:5.2f} ({:5.2f})'
+        residuals_contribution_format_str = '{:' + str(longest_name_len + 4) + '}{:5.2f}'
+
+        # add fits in descending order of reference contribution
+        reference_line_list = []
+        reference_label_list = []
+        reference_contributions_percent_sr.sort_values(ascending=False, inplace=True)
+        reference_only_contributions_percent_sr.sort_values(ascending=False, inplace=True)
+        log.info('plotting reference components')
+        log.info(reference_contributions_percent_sr.head())
+        for (ref_name, ref_contrib), (ref_only_name, ref_only_contrib) \
+                in zip(reference_contributions_percent_sr.items(), reference_only_contributions_percent_sr.items()):
+            log.info('reference contribution {} {}'.format(ref_name, ref_contrib))
+            log.info('reference-only contribution {} {}'.format(ref_only_name, ref_only_contrib))
+            reference_label = reference_contribution_format_str.format(ref_name, ref_contrib, ref_only_contrib)
+            reference_label_list.append(reference_label)
+
+            # plot once for each reference just to build the legend
+            # ax.plot returns a list
+            reference_line_list.extend(
+                ax.plot(
+                    any_given_fit.interpolant_incident_energy,
+                    any_given_fit.fit_spectrum_b,
+                    label=reference_label,
+                    color='w',
+                    alpha=0.0))
+
+        log.info(any_given_fit.residuals.shape)
+        residuals_label = residuals_contribution_format_str.format('residuals', any_given_fit.residuals_contribution)
+        residuals_line = ax.plot(
+            any_given_fit.interpolant_incident_energy,
+            any_given_fit.residuals,
+            label=residuals_label)
+
+        fit_line_label = 'fit'
+        fit_line = ax.plot(
+            any_given_fit.interpolant_incident_energy,
+            any_given_fit.fit_spectrum_b,
+            label=fit_line_label)
+
+        spectrum_points = ax.plot(
+            any_given_fit.interpolant_incident_energy,
+            any_given_fit.unknown_spectrum_b,
+            '.',
+            label=spectrum.file_name,
+            alpha=0.5)
+
+        ax.set_xlabel('eV')
+        ax.set_ylabel('normalized absorbance')
+        # 20171029
+        ax.legend(
+            [*reference_line_list, *spectrum_points, *residuals_line, *fit_line],
+            [*reference_label_list, spectrum.file_name, residuals_label, fit_line_label],
+            prop=dict(family='Monospace', size=7))
+
+        return f
+
+    def plot_stacked_fit(self, spectrum, any_given_fit, title):
+        log = logging.getLogger(name=self.__class__.__name__)
+
+        f, ax = plt.subplots()
+        f.suptitle(spectrum.file_name + '\n' + title)
+        log.info(any_given_fit.fit_spectrum_b.shape)
+
+        reference_contributions_percent_sr = any_given_fit.get_reference_contributions_sr()
+        longest_name_len = max([len(name) for name in reference_contributions_percent_sr.index] + [len(spectrum.file_name)])
+        # the format string should look like '{:N}{:5.2f}' where N is the length of the longest reference name
+        contribution_format_str = '{:' + str(longest_name_len + 4) + '}{:5.2f}'
+
+        log.info(any_given_fit.residuals.shape)
+        residuals_label = contribution_format_str.format('residuals', any_given_fit.residuals_contribution)
+        residuals_line = ax.plot(
+            any_given_fit.interpolant_incident_energy,
+            any_given_fit.residuals,
+            label=residuals_label)
+
+        spectrum_points = ax.plot(
+            any_given_fit.interpolant_incident_energy,
+            any_given_fit.unknown_spectrum_b,
+            '.',
+            label=spectrum.file_name,
+            alpha=0.5)
+
+        # add fits in descending order of reference contribution
+        reference_label_list = []
+        reference_contributions_percent_sr.sort_values(ascending=False, inplace=True)
+        sort_ndx = reversed(any_given_fit.reference_spectra_coef_x.argsort())
+        ys = any_given_fit.reference_spectra_coef_x * any_given_fit.reference_spectra_A_df
+        log.info('plotting reference components')
+        log.info(reference_contributions_percent_sr.head())
+        reference_contributions_percent_sr.sort_values(ascending=False)
+        for name, value in reference_contributions_percent_sr.items():
+            log.info('reference component {} {}'.format(name, value))
+            reference_label = contribution_format_str.format(name, value)
+            reference_label_list.append(reference_label)
+
+        reference_line_list = ax.stackplot(ys.index, *[ys.iloc[:, i] for i in sort_ndx], labels=reference_label_list)
+
+        ax.set_xlabel('eV')
+        ax.set_ylabel('normalized absorbance')
+        ax.legend(
+            # these arguments are documented but this does not seem to work
+            [*spectrum_points, *reference_line_list, *residuals_line],
+            [spectrum.file_name, *reference_label_list, residuals_label],
+            prop=dict(family='Monospace', size=7))
+
+        return f
 
     def draw_plots_bokeh(self, plots_html_file_path):
         log = logging.getLogger(name=self.__class__.__name__)
@@ -357,6 +429,34 @@ class AllCombinationFitTask:
         p = bokeh.models.layouts.Column(*plot_list)
         bokeh.io.save(p)
 
+    def bokeh_nss_path(self, spectrum, fit_results, output_fp):
+        log = logging.getLogger(name=self.__class__.__name__)
+        bokeh.plotting.output_file(output_fp)
+
+        hover = bokeh.models.HoverTool(tooltips=[
+            ("index", "$index"),
+            ("NSS", "$y"),
+            ("desc", "@desc"),
+        ])
+        p = bokeh.plotting.figure(
+            plot_width=400, plot_height=400,
+            tools=['reset', hover, 'box_zoom'],
+            title=spectrum.file_name,
+            x_axis_label='index',
+            y_axis_label='NSS')
+
+        for component_count in fit_results.component_count_fit_table.keys():
+            sorted_fits = fit_results.component_count_fit_table[component_count][:20]
+            source = bokeh.plotting.ColumnDataSource(data=dict(
+                x=range(len(sorted_fits)),
+                nss=[spectrum_fit.nss for spectrum_fit in sorted_fits],
+                desc=[str(spectrum_fit) for spectrum_fit in sorted_fits]
+            ))
+
+            p.circle('x', 'nss', legend='{}-component'.format(component_count), size=5, source=source)
+
+        bokeh.io.save(p)
+
     def write_best_fit_arrays(self, best_fit_dir_path):
         log = logging.getLogger(name=self.__class__.__name__)
         for spectrum, fit_results in self.fit_table.items():
@@ -372,144 +472,3 @@ class AllCombinationFitTask:
                     'residual': fit_results.best_fit.residuals}
             )
             fit_df.to_csv(fit_file_path, sep='\t', float_format='%8.4f', index=False)
-
-    @classmethod
-    def build_reference_spectrum_list_from_prm_file(cls, prm_file_path):
-        """
-        Read a PRM file to create a list of ReferenceSpectrum
-        instances, maximum component count, and minimum component
-        count from a PRM file.
-
-        :param prm_file_path:
-        :return:
-            list of ReferenceSpectrum instances
-            maximum component count
-            minimum component count
-        """
-        log = logging.getLogger(name=cls.__class__.__name__)
-        reference_spectrum_list = []
-        log.info('reading PRM file {}'.format(prm_file_path))
-        prm = PRM.read_prm(prm_file_path)
-        # read reference files
-        for i, fp in enumerate(prm.reference_file_path_list):
-            log.info('reading reference file {}: {}'.format(i, fp))
-            reference_spectrum = ReferenceSpectrum.read_file(fp)
-            reference_spectrum_list.append(reference_spectrum)
-
-        return reference_spectrum_list, prm.nb_component_max, prm.nb_component_min
-
-    @classmethod
-    def build_reference_spectrum_list_from_config_file(cls, config):
-        """
-        Read reference spectrum file glob(s) from configuration file to create
-        a list of ReferenceSpectrum instances, maximum component count, and
-        minimum component count.
-
-        :param config: configparser instance
-        :return: list of ReferenceSpectrum instances
-        """
-
-        reference_spectrum_list, reference_spectrum_file_glob_list = ReferenceSpectrum.read_all(
-            [reference_file_glob for reference_file_glob, _ in config.items('references')]
-        )
-
-        if len(reference_spectrum_file_glob_list) == 0:
-            logging.exception(
-                'no reference spectrum file paths or patterns were found in section [references] of configuration file {}'.format(
-                    config
-                )
-            )
-        elif len(reference_spectrum_list) == 0:
-            logging.exception('no reference spectrum files were found')
-        else:
-            # everything is ok
-            pass
-
-        if not config.has_option('fit', 'maximum_component_count'):
-            logging.exception(
-                'required parameter maximum_component_count is missing from section [fit] in configuration file {}'.format(
-                    config
-                )
-            )
-        elif not config.has_option('fit', 'minimum_component_count'):
-            logging.exception(
-                'required parameter minimum_component_count is missing from section [fit] in configuration file {}'.format(
-                    config
-                )
-            )
-        else:
-            maximum_component_count = config.getint('fit', 'maximum_component_count')
-            minimum_component_count = config.getint('fit', 'minimum_component_count')
-
-        return reference_spectrum_list, maximum_component_count, minimum_component_count
-
-    @classmethod
-    def build(cls, config):
-        log = logging.getLogger(name=str(cls))
-
-        # read section [references]
-        # support a PRM file such as
-        #   prm = path/to/one.prm
-        # or
-        # a list of one or more file globs such as
-        #   arsenic_2_reference_spectra/*.e
-        #   arsenic_3_reference_spectra/*.e
-
-        if config.has_section('references'):
-            if config.has_option('references', 'prm'):
-                prm_file_path = os.path.expanduser(config.get('references', 'prm'))
-                reference_spectrum_list, max_cmp, min_cmp = cls.build_reference_spectrum_list_from_prm_file(prm_file_path)
-            else:
-                reference_spectrum_list, max_cmp, min_cmp = cls.build_reference_spectrum_list_from_config_file(config)
-        elif config.has_section('reference_spectra'):
-            if config.has_option('reference_spectra', 'prm'):
-                prm_file_path = os.path.expanduser(config.get('reference_spectra', 'prm'))
-                reference_spectrum_list, max_cmp, min_cmp = cls.build_reference_spectrum_list_from_prm_file(prm_file_path)
-            else:
-                raise Exception('section [reference_spectra] is missing required parameter prm')
-        else:
-            raise Exception('configuration file is missing section [references]')
-
-        if 0 < min_cmp <= max_cmp:
-            component_count_range = range(min_cmp, max_cmp+1)
-            logging.info('component count range: {}'.format(component_count_range))
-        else:
-            logging.exception('minimum and maximum component counts are not valid')
-
-        energy_range = cls.get_energy_range_from_config(config)
-
-        # read data files
-        unknown_spectrum_file_path_list = []
-        for j, (unknown_spectrum_glob, _) in enumerate(config.items('data')):
-            log.info('unknown spectrum glob: {}'.format(unknown_spectrum_glob))
-            glob_pattern_expanded = os.path.expanduser(unknown_spectrum_glob)
-            unknown_spectrum_file_path_list.extend(glob.glob(glob_pattern_expanded))
-        log.info('found {} data files'.format(len(unknown_spectrum_file_path_list)))
-
-        unknown_spectrum_list = []
-        for unknown_spectrum_file_path in unknown_spectrum_file_path_list:
-            log.info('reading data file {}'.format(unknown_spectrum_file_path))
-            unknown_spectrum = Spectrum.read_file(unknown_spectrum_file_path)
-            unknown_spectrum_list.append(unknown_spectrum)
-
-        fit_task = cls(
-            reference_spectrum_list=reference_spectrum_list,
-            unknown_spectrum_list=unknown_spectrum_list,
-            energy_range_builder=energy_range,
-            component_count_range=component_count_range
-        )
-
-        return fit_task
-
-    @classmethod
-    def get_energy_range_from_config(cls, config):
-        if config.has_option('parameters', 'fit_energy_start') and config.has_option('parameters', 'fit_energy_stop'):
-            fit_energy_start = config.getfloat('parameters', 'fit_energy_start')
-            fit_energy_stop = config.getfloat('parameters', 'fit_energy_stop')
-            energy_range = FixedEnergyRangeBuilder(fit_energy_start, fit_energy_stop)
-        elif not(config.has_option('parameters', 'fit_energy_start') or config.has_option('parameters', 'fit_energy_stop')):
-            energy_range = AdaptiveEnergyRangeBuilder()
-        else:
-            raise Exception('only one of fit_energy_start and fit_energy_stop was specified in the configuration')
-
-        return energy_range